[BUGFIX] adds colPos to showitems of all registered content elements
In a blank project for 7.6 there is no colPos definition for the custom contentelements created in dw_content_elements_source, therefore we have to add it permanently.
This should be done in dw_content_elements, because it affects all custom content elements, otherwise we have to add it in the content element typoscript configuration.
If we don't have colPos defined, we are not able to create new custom content elements in an other column than 0.